proyecto_app_movil/backend/requirements.txt
python-dotenv
openai
flask
flask-cors
gunicorn